requirements.txt
flask==3.1.2
PyJWT==2.10.1
PyYAML==6.0.3
python-dotenv==1.1.1
Werkzeug==3.1.3
watchdog==6.0.0
requests==2.32.5
